feat(api): 优化用户接口和分页基类
- PageParam 改为 class 支持继承，加排序字段和不分页支持
- PageResult 加 @Schema 注解
- UserPageDTO 继承 PageParam，去除重复字段
- UserController/AuthController 加 @Tag + @Operation 注解
- 新增批量删除、修改状态接口
- create 接口返回 201，resetPassword 改为 POST
- status 字段统一为 Integer，修复测试

+++ apps/service/aaf-api/src/main/java/com/xuejiai/aaf/module/system/vo/UserVO.java
@@ -7,6 +7,6 @@ public record UserVO(
         Long id,
         String username,
         String nickname,
-        Short status,
+        Integer status,
         LocalDateTime createTime,
         LocalDateTime updateTime) {}
